api_yamdb/requirements.txt
asgiref==3.2.10
django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
PyJWT==2.1.0
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
psycopg2-binary==2.9.5
flake8==4.0.1
pep8-naming==0.13.2
flake8-broken-line==0.6.0
flake8-return==1.2.0
flake8-isort==4.1.2